Remove model_definition.py
